+++ Magic/src/main/java/com/elmakers/mine/bukkit/resourcepack/ResourcePackResponse.java
@@ -3,5 +3,5 @@ package com.elmakers.mine.bukkit.resourcepack;
 import java.util.List;
 
 public interface ResourcePackResponse {
-    void finished(boolean success, List<String> responses, ResourcePack pack);
+    void finished(boolean success, boolean hasModifiedTime, List<String> responses, ResourcePack pack);
 }