+++ Magic/CHANGELOG.md
@@ -53,6 +53,7 @@
  - Shop and Selector default messages are now different, base selector messages may need localization
  - Default wand item flags changed to hide dye color. Add an `item_flags` list to your wand config if you don't want that.
  - Wand charges functionality changed to work more like per-spell mana
+ - Improvements to alternate and per-world resource pack functionality
 
 # 9.3